node-prep.sh: make the extraction of the reloc package reentrant

+++ node_prep.sh
@@ -16,10 +16,16 @@ sudo curl -o /etc/yum.repos.d/scylla.repo -L http://downloads.scylladb.com.s3.am
 
 sudo yum install tmux scylla-gdb scylla-tools scylla-jmx scylla -y
 
-mkdir scylla-relocatable-package
-mv ./scylla-relocatable-package.tar.gz scylla-relocatable-package
-cd scylla-relocatable-package
-tar -xf ./scylla-relocatable-package.tar.gz
+if [ ! -e scylla-relocatable-package ]
+then
+    mkdir scylla-relocatable-package
+    mv ./scylla-relocatable-package.tar.gz scylla-relocatable-package
+    cd scylla-relocatable-package
+    tar -xf ./scylla-relocatable-package.tar.gz
+else
+    cd scylla-relocatable-package
+fi
+
 sudo ./install.sh
 
 sudo sed -i "s/#cluster_name: '.*Cluster'/cluster_name: '${CLUSTER_NAME}'/" ${SCYLLA_YAML}
